requirements.txt
alembic==1.8.1
autopep8==2.0.0
Babel==2.11.0
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
docutils==0.19
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
pycodestyle==2.10.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
